requirements.txt
streamlit
pandas
openpyxl
pdfplumber
python-docx